extract TodoItem and TodoList components
