+++ src/App.tsx
@@ -6,6 +6,7 @@ import classNames from 'classnames';
 import { UserWarning } from './UserWarning';
 import { USER_ID, addTodo, deleteTodo, getTodos } from './api/todos';
 import { Todo } from './types/Todo';
+import { TodoList } from './components/TodoList';
 
 type FilterStatus = 'all' | 'active' | 'completed';
 
@@ -227,53 +228,11 @@ export const App: React.FC = () => {
           })}
           data-cy="TodoList"
         >
-          {filteredTodos.map(todo => {
-            const isDeleting = deletingIds.includes(todo.id);
-
-            return (
-              <div
-                key={todo.id}
-                data-cy="Todo"
-                className={classNames('todo', { completed: todo.completed })}
-              >
-                <label className="todo__status-label">
-                  <input
-                    data-cy="TodoStatus"
-                    type="checkbox"
-                    className="todo__status"
-                    checked={todo.completed}
-                    readOnly
-                  />
-                </label>
-
-                <span data-cy="TodoTitle" className="todo__title">
-                  {todo.title}
-                </span>
-
-                {/* Remove button appears only on hover */}
-                <button
-                  type="button"
-                  className="todo__remove"
-                  data-cy="TodoDelete"
-                  onClick={() => handleDeleteTodo(todo.id)}
-                  disabled={isDeleting}
-                >
-                  ×
-                </button>
-
-                {/* overlay will cover the todo while it is being deleted or updated */}
-                <div
-                  data-cy="TodoLoader"
-                  className={classNames('modal overlay', {
-                    'is-active': isDeleting,
-                  })}
-                >
-                  <div className="modal-background has-background-white-ter" />
-                  <div className="loader" />
-                </div>
-              </div>
-            );
-          })}
+          <TodoList
+            todos={filteredTodos}
+            deletingIds={deletingIds}
+            onDelete={handleDeleteTodo}
+          />
 
           {tempTodo && (
             <div data-cy="Todo" className="todo">
